Moving pathSplitter and makeNewDataSnap to separate helpers file, and exporting makeNewDataSnap in test-helpers
