+++ lib/ruleset.js
@@ -3,7 +3,9 @@
 
 var extend = require('extend'),
   Rule = require('./parser/rule'),
-  RuleDataSnapshot = require('./rule-data-snapshot');
+  RuleDataSnapshot = require('./rule-data-snapshot'),
+  pathSplitter = require('./helpers').pathSplitter,
+  makeNewDataSnap = require('./helpers').makeNewDataSnap;
 
 var validRuleKinds = {
   '.read': true,
@@ -19,47 +21,6 @@ function ruleError(rulePath, message) {
   return err;
 
 }
-
-function pathSplitter(path)
-{
-
-  if (path[0] === '/') {
-    path = path.substr(1);
-  }
-
-  return path.split('/');
-
-}
-
-function makeNewDataSnap(path, newData) {
-
-  path = path.replace(/^\/+/, '');
-  var result;
-
-  if (path.length === 0) {
-    result = newData;
-  } else {
-
-    var workObject = {};
-    result = workObject;
-
-    pathSplitter(path).forEach(function(pathPart, i, pathParts) {
-
-      if (pathParts.length - i === 1) {
-        workObject[pathPart] = newData;
-      } else {
-        workObject[pathPart] = {};
-      }
-      workObject = workObject[pathPart];
-
-    });
-
-  }
-
-  return new RuleDataSnapshot(result);
-
-}
-
 
 function Ruleset(rulesDefinition) {
 
@@ -152,7 +113,6 @@ function Ruleset(rulesDefinition) {
   this._rules = newRulesObject;
 
 }
-
 
 Ruleset.prototype.get = function(path, kind) {
