+++ lib/test-helpers.js
@@ -4,7 +4,8 @@
 var root, rules,
   debug = true,
   Ruleset = require('./ruleset'),
-  RuleDataSnapshot = require('./rule-data-snapshot');
+  RuleDataSnapshot = require('./rule-data-snapshot'),
+  makeNewDataSnap = require('./helpers').makeNewDataSnap;
 
 var userDefinitions = exports.userDefinitions = {
 
@@ -165,3 +166,5 @@ exports.setFirebaseRules = function(ruleDefinition) {
 exports.getFirebaseRules = function() {
   return rules;
 };
+
+exports.makeNewDataSnap = makeNewDataSnap;
